pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maxagent"
version = "0.1.0"
description = "AI Code Assistant CLI based on LiteLLM + GitHub Copilot"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "maxazure", email = "maxazure@gmail.com" }
]
keywords = ["cli", "ai", "code-assistant", "llm", "copilot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.0",
]
web = [
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]
all = [
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]

[project.scripts]
llc = "maxagent.cli.main:cli"
maxagent = "maxagent.cli.main:cli"

[project.urls]
Homepage = "https://github.com/maxazure/maxagent"
Repository = "https://github.com/maxazure/maxagent"

[tool.hatch.build.targets.wheel]
packages = ["src/maxagent"]

[tool.ruff]
target-version = "py312"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["maxagent"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src/maxagent --cov-report=term-missing"
